requirements.txt
semaphore-bot
openai